[#129] refactor: EnchantRankTable 컴포넌트 커스텀 훅으로 분리
