+++ .gitignore
@@ -44,3 +44,4 @@ dummyEnchantList.ts
 src/app/_components/testComponents/*
 src/app/(route)/(main)/testpage/*
 delete/*
+/src/app/_components/character/*
